Removed version pinning
The version pinning on Nielsen DTVR is blocking Fox from updating to the latest ios library (4.0.1). which is necessary for them to resolve other open issues. It also restricts them from using new features like destinationMiddlewares in analytics-ios 4.x. This is high urgency for Fox. Please let me know if any testing is needed to merge this and I can do so!